v09
Prior to this revision, the toggle button simply toggles between play and pause.  However this does not allow for mpd status of "stop" or "unknown" (e.g. after a reboot).  Now when the button is pressed, if mpd state is "stop" or "unknown" then the daemon sends the play command.  If mpd status is "play" or "pause" then the daemon sends the "pause_toggle" command.